feat: 보드 및 컴포넌트 목록 인터페이스 구현
