--- src/main/kotlin/minesweeper/domain/Board.kt
@@ -1,7 +1,0 @@
-package minesweeper.domain
-
-class Board(
-    private val height: Int,
-    private val width: Int,
-    private val mineCount: Int
-)
