+++ src/main/kotlin/minesweeper/domain/board/Board.kt
@@ -1,0 +1,8 @@
+package minesweeper.domain.board
+
+import minesweeper.domain.component.Components
+
+interface Board : Components {
+    val height: Int
+    val width: Int
+}
