+++ src/main/kotlin/minesweeper/domain/component/Components.kt
@@ -1,0 +1,5 @@
+package minesweeper.domain.component
+
+interface Components {
+    fun components(): List<Component>
+}
